Reorganize install tests
