+++ tests/unit/install/conftest.py
@@ -1,0 +1,102 @@
+import io
+from unittest.mock import create_autospec
+
+import pytest
+import yaml
+from databricks.labs.blueprint.tui import MockPrompts
+from databricks.sdk import WorkspaceClient
+from databricks.sdk.service.compute import ClusterDetails, CreatePolicyResponse, DataSecurityMode, State
+from databricks.sdk.errors import NotFound
+from databricks.sdk.service import iam, jobs
+from databricks.sdk.service.sql import (
+    Dashboard,
+    DataSource,
+    EndpointInfo,
+    EndpointInfoWarehouseType,
+    Query,
+    Visualization,
+    Widget,
+)
+from databricks.sdk.service.workspace import ObjectInfo
+
+
+@pytest.fixture
+def any_prompt():
+    return MockPrompts({".*": ""})
+
+
+@pytest.fixture
+def clusters() -> list[ClusterDetails]:
+    return [
+        ClusterDetails(
+            spark_version="13.3.x-dbrxxx",
+            cluster_name="zero",
+            data_security_mode=DataSecurityMode.USER_ISOLATION,
+            state=State.RUNNING,
+            cluster_id="1111-999999-userisol",
+        ),
+        ClusterDetails(
+            spark_version="13.3.x-dbrxxx",
+            cluster_name="one",
+            data_security_mode=DataSecurityMode.NONE,
+            state=State.RUNNING,
+            cluster_id='2222-999999-nosecuri',
+        ),
+        ClusterDetails(
+            spark_version="13.3.x-dbrxxx",
+            cluster_name="two",
+            data_security_mode=DataSecurityMode.LEGACY_TABLE_ACL,
+            state=State.RUNNING,
+            cluster_id='3333-999999-legacytc',
+        ),
+    ]
+
+
+@pytest.fixture
+def ws(clusters):
+    state = {
+        "/Applications/ucx/config.yml": yaml.dump(
+            {
+                'version': 1,
+                'inventory_database': 'ucx_exists',
+                'connect': {
+                    'host': '...',
+                    'token': '...',
+                },
+            }
+        ),
+    }
+
+    def download(path: str) -> io.StringIO | io.BytesIO:
+        if path not in state:
+            raise NotFound(path)
+        if ".csv" in path:
+            return io.BytesIO(state[path].encode('utf-8'))
+        return io.StringIO(state[path])
+
+    workspace_client = create_autospec(WorkspaceClient)
+
+    workspace_client.current_user.me = lambda: iam.User(
+        user_name="me@example.com", groups=[iam.ComplexValue(display="admins")]
+    )
+    workspace_client.config.host = "https://foo"
+    workspace_client.config.is_aws = True
+    workspace_client.config.is_azure = False
+    workspace_client.config.is_gcp = False
+    workspace_client.workspace.get_status = lambda _: ObjectInfo(object_id=123)
+    workspace_client.data_sources.list = lambda: [DataSource(id="bcd", warehouse_id="abc")]
+    workspace_client.warehouses.list = lambda **_: [
+        EndpointInfo(name="abc", id="abc", warehouse_type=EndpointInfoWarehouseType.PRO, state=State.RUNNING)
+    ]
+    workspace_client.dashboards.create.return_value = Dashboard(id="abc")
+    workspace_client.jobs.create.return_value = jobs.CreateResponse(job_id=123)
+    workspace_client.queries.create.return_value = Query(id="abc")
+    workspace_client.query_visualizations.create.return_value = Visualization(id="abc")
+    workspace_client.dashboard_widgets.create.return_value = Widget(id="abc")
+    workspace_client.clusters.list.return_value = clusters
+    workspace_client.cluster_policies.create.return_value = CreatePolicyResponse(policy_id="foo")
+    workspace_client.clusters.select_spark_version = lambda **_: "14.2.x-scala2.12"
+    workspace_client.clusters.select_node_type = lambda **_: "Standard_F4s"
+    workspace_client.workspace.download = download
+
+    return workspace_client
