+++ tests/unit/install/test_install.py
@@ -1,4 +1,3 @@
-import io
 import json
 import logging
 from datetime import datetime, timedelta
@@ -6,15 +5,13 @@ from typing import Any
 from unittest.mock import MagicMock, create_autospec, patch
 
 import pytest
-import yaml
 from requests.exceptions import ConnectionError as RequestsConnectionError
 from databricks.labs.blueprint.installation import Installation, MockInstallation
 from databricks.labs.blueprint.installer import InstallState
 from databricks.labs.blueprint.parallel import ManyError
 from databricks.labs.blueprint.tui import MockPrompts
-from databricks.labs.blueprint.wheels import ProductInfo, WheelsV2, find_project_root
+from databricks.labs.blueprint.wheels import ProductInfo, WheelsV2
 from databricks.labs.lsql.backends import MockBackend
-from databricks.labs.lsql.dashboards import DashboardMetadata
 from databricks.sdk import AccountClient, WorkspaceClient
 from databricks.sdk.core import Config
 from databricks.sdk.credentials_provider import credentials_strategy
@@ -29,20 +26,10 @@ from databricks.sdk.errors import (  # pylint: disable=redefined-builtin
 )
 from databricks.sdk.errors.platform import BadRequest
 from databricks.sdk.service import iam, jobs, sql
-from databricks.sdk.service.compute import ClusterDetails, CreatePolicyResponse, DataSecurityMode, Policy, State
+from databricks.sdk.service.compute import Policy, State
 from databricks.sdk.service.jobs import BaseRun, RunLifeCycleState, RunResultState, RunState
 from databricks.sdk.service.provisioning import Workspace
-from databricks.sdk.service.dashboards import Dashboard as LakeviewDashboard, LifecycleState
-from databricks.sdk.service.sql import (
-    Dashboard,
-    DataSource,
-    EndpointInfo,
-    EndpointInfoWarehouseType,
-    Query,
-    Visualization,
-    Widget,
-)
-from databricks.sdk.service.workspace import ObjectInfo
+from databricks.sdk.service.sql import EndpointInfo, EndpointInfoWarehouseType
 
 import databricks.labs.ucx.installer.mixins
 import databricks.labs.ucx.uninstall  # noqa
@@ -52,82 +39,6 @@ from databricks.labs.ucx.installer.workflows import DeployedWorkflows, Workflows
 from databricks.labs.ucx.runtime import Workflows
 
 PRODUCT_INFO = ProductInfo.from_class(WorkspaceConfig)
-
-
-def mock_clusters():
-    return [
-        ClusterDetails(
-            spark_version="13.3.x-dbrxxx",
-            cluster_name="zero",
-            data_security_mode=DataSecurityMode.USER_ISOLATION,
-            state=State.RUNNING,
-            cluster_id="1111-999999-userisol",
-        ),
-        ClusterDetails(
-            spark_version="13.3.x-dbrxxx",
-            cluster_name="one",
-            data_security_mode=DataSecurityMode.NONE,
-            state=State.RUNNING,
-            cluster_id='2222-999999-nosecuri',
-        ),
-        ClusterDetails(
-            spark_version="13.3.x-dbrxxx",
-            cluster_name="two",
-            data_security_mode=DataSecurityMode.LEGACY_TABLE_ACL,
-            state=State.RUNNING,
-            cluster_id='3333-999999-legacytc',
-        ),
-    ]
-
-
-@pytest.fixture
-def ws():
-    state = {
-        "/Applications/ucx/config.yml": yaml.dump(
-            {
-                'version': 1,
-                'inventory_database': 'ucx_exists',
-                'connect': {
-                    'host': '...',
-                    'token': '...',
-                },
-            }
-        ),
-    }
-
-    def download(path: str) -> io.StringIO | io.BytesIO:
-        if path not in state:
-            raise NotFound(path)
-        if ".csv" in path:
-            return io.BytesIO(state[path].encode('utf-8'))
-        return io.StringIO(state[path])
-
-    workspace_client = create_autospec(WorkspaceClient)
-
-    workspace_client.current_user.me = lambda: iam.User(
-        user_name="me@example.com", groups=[iam.ComplexValue(display="admins")]
-    )
-    workspace_client.config.host = "https://foo"
-    workspace_client.config.is_aws = True
-    workspace_client.config.is_azure = False
-    workspace_client.config.is_gcp = False
-    workspace_client.workspace.get_status = lambda _: ObjectInfo(object_id=123)
-    workspace_client.data_sources.list = lambda: [DataSource(id="bcd", warehouse_id="abc")]
-    workspace_client.warehouses.list = lambda **_: [
-        EndpointInfo(name="abc", id="abc", warehouse_type=EndpointInfoWarehouseType.PRO, state=State.RUNNING)
-    ]
-    workspace_client.dashboards.create.return_value = Dashboard(id="abc")
-    workspace_client.jobs.create.return_value = jobs.CreateResponse(job_id=123)
-    workspace_client.queries.create.return_value = Query(id="abc")
-    workspace_client.query_visualizations.create.return_value = Visualization(id="abc")
-    workspace_client.dashboard_widgets.create.return_value = Widget(id="abc")
-    workspace_client.clusters.list.return_value = mock_clusters()
-    workspace_client.cluster_policies.create.return_value = CreatePolicyResponse(policy_id="foo")
-    workspace_client.clusters.select_spark_version = lambda **_: "14.2.x-scala2.12"
-    workspace_client.clusters.select_node_type = lambda **_: "Standard_F4s"
-    workspace_client.workspace.download = download
-
-    return workspace_client
 
 
 def created_job(workspace_client, name):
@@ -181,11 +92,6 @@ def mock_installation_extra_jobs():
             }
         }
     )
-
-
-@pytest.fixture
-def any_prompt():
-    return MockPrompts({".*": ""})
 
 
 def not_found(_):
@@ -619,162 +525,6 @@ def test_main_with_existing_conf_does_not_recreate_config(ws, mocker, mock_insta
     webbrowser_open.assert_called_with('https://localhost/#workspace~/mock/README')
     wheels.upload_to_wsfs.assert_called_once()
     wheels.upload_to_dbfs.assert_not_called()
-
-
-@pytest.fixture
-def new_workspace_installation(request, ws, any_prompt) -> WorkspaceInstallation:
-    mock_installation = request.param if hasattr(request, "param") else MockInstallation()
-    install_state = InstallState.from_installation(mock_installation)
-    wheels = WheelsV2(mock_installation, PRODUCT_INFO)
-    workflows_installation = WorkflowsDeployment(
-        WorkspaceConfig(inventory_database="...", policy_id='123'),
-        mock_installation,
-        install_state,
-        ws,
-        wheels,
-        PRODUCT_INFO,
-        timedelta(seconds=1),
-        [],
-    )
-    workspace_installation = WorkspaceInstallation(
-        WorkspaceConfig(inventory_database='ucx'),
-        mock_installation,
-        install_state,
-        MockBackend(),
-        ws,
-        workflows_installation,
-        any_prompt,
-        PRODUCT_INFO,
-    )
-    return workspace_installation
-
-
-def test_installation_creates_dashboard(ws, new_workspace_installation):
-    new_workspace_installation.run()
-    ws.lakeview.create.assert_called()
-
-
-def test_installation_updates_dashboard(ws, new_workspace_installation):
-    new_workspace_installation.run()
-    ws.lakeview.update.assert_not_called()
-    new_workspace_installation.run()
-    ws.lakeview.update.assert_called()
-
-
-@pytest.mark.parametrize(
-    "new_workspace_installation",
-    [MockInstallation({'state.json': {'resources': {'dashboards': {"assessment_main": "redash-id"}}}})],
-    indirect=True,
-)
-def test_installation_upgrades_redash_dashboard(caplog, ws, new_workspace_installation):
-    with caplog.at_level(logging.INFO, logger="databricks.labs.ucx.install"):
-        new_workspace_installation.run()
-    assert "Upgrading dashboard to Lakeview" in caplog.text
-    ws.dashboards.delete.assert_called_once()
-
-
-@pytest.mark.parametrize(
-    "new_workspace_installation",
-    [MockInstallation({'state.json': {'resources': {'dashboards': {"assessment_main": "redash-id"}}}})],
-    indirect=True,
-)
-def test_installation_upgrades_non_existing_redash_dashboard(caplog, ws, new_workspace_installation):
-    ws.dashboards.delete.side_effect = BadRequest
-    with caplog.at_level(logging.INFO, logger="databricks.labs.ucx.install"):
-        new_workspace_installation.run()
-    assert "Upgrading dashboard to Lakeview" in caplog.text
-    assert "Cannot delete dashboard" in caplog.text
-    ws.dashboards.delete.assert_called_once()
-
-
-@pytest.mark.parametrize(
-    "new_workspace_installation",
-    [MockInstallation({'state.json': {'resources': {'dashboards': {"assessment_main": "lakeview_id"}}}})],
-    indirect=True,
-)
-def test_installation_updates_existing_lakeview_dashboard(ws, new_workspace_installation):
-    ws.lakeview.get.return_value = LakeviewDashboard(lifecycle_state=LifecycleState.ACTIVE)
-    new_workspace_installation.run()
-    ws.lakeview.get.assert_called_with("lakeview_id")
-    ws.lakeview.update.assert_called_once()
-
-
-@pytest.mark.parametrize(
-    "new_workspace_installation",
-    [MockInstallation({'state.json': {'resources': {'dashboards': {"assessment_main": "lakeview_id"}}}})],
-    indirect=True,
-)
-def test_installation_recreates_lakeview_dashboard_without_lifecycle_state(caplog, ws, new_workspace_installation):
-    ws.lakeview.get.return_value = LakeviewDashboard(lifecycle_state=None)
-    with caplog.at_level(logging.INFO, logger="databricks.labs.ucx.install"):
-        new_workspace_installation.run()
-    assert "Recovering invalid dashboard" in caplog.text
-    ws.lakeview.create.assert_called()
-    ws.lakeview.update.assert_not_called()
-
-
-@pytest.mark.parametrize(
-    "new_workspace_installation",
-    [MockInstallation({'state.json': {'resources': {'dashboards': {"assessment_main": "lakeview_id"}}}})],
-    indirect=True,
-)
-def test_installation_recreates_trashed_lakeview_dashboard(caplog, ws, new_workspace_installation):
-    ws.lakeview.get.return_value = LakeviewDashboard(lifecycle_state=LifecycleState.TRASHED)
-    with caplog.at_level(logging.INFO, logger="databricks.labs.ucx.install"):
-        new_workspace_installation.run()
-    assert "Recreating trashed dashboard" in caplog.text
-    ws.lakeview.create.assert_called()
-    ws.lakeview.update.assert_not_called()
-
-
-@pytest.mark.parametrize(
-    "new_workspace_installation",
-    [MockInstallation({'state.json': {'resources': {'dashboards': {"assessment_main": "lakeview_id"}}}})],
-    indirect=True,
-)
-@pytest.mark.parametrize("exception", [NotFound, InvalidParameterValue])
-def test_installation_recovers_invalid_dashboard(caplog, ws, new_workspace_installation, exception):
-    ws.lakeview.get.side_effect = exception
-    with caplog.at_level(logging.DEBUG, logger="databricks.labs.ucx.install"):
-        new_workspace_installation.run()
-    assert "Recovering invalid dashboard" in caplog.text
-    assert "Deleted dangling dashboard" in caplog.text
-    ws.workspace.delete.assert_called_once()
-    ws.lakeview.create.assert_called()
-    ws.lakeview.update.assert_not_called()
-
-
-@pytest.mark.parametrize(
-    "new_workspace_installation",
-    [MockInstallation({'state.json': {'resources': {'dashboards': {"assessment_main": "lakeview_id"}}}})],
-    indirect=True,
-)
-def test_installation_recovers_invalid_deleted_dashboard(caplog, ws, new_workspace_installation):
-    ws.lakeview.get.side_effect = NotFound
-    ws.workspace.delete.side_effect = NotFound
-    with caplog.at_level(logging.DEBUG, logger="databricks.labs.ucx.install"):
-        new_workspace_installation.run()
-    assert "Recovering invalid dashboard" in caplog.text
-    assert "Deleted dangling dashboard" not in caplog.text
-    ws.workspace.delete.assert_called_once()
-    ws.lakeview.create.assert_called()
-    ws.lakeview.update.assert_not_called()
-
-
-def test_validate_dashboards(ws):
-    queries_path = find_project_root(__file__) / "src/databricks/labs/ucx/queries"
-    for step_folder in queries_path.iterdir():
-        if not step_folder.is_dir():
-            continue
-        for dashboard_folder in step_folder.iterdir():
-            if not dashboard_folder.is_dir():
-                continue
-            try:
-                DashboardMetadata.from_path(dashboard_folder).validate()
-            except ValueError as e:
-                assert False, f"Invalid dashboard in {dashboard_folder}: {e}"
-            else:
-                assert True, f"Valid dashboard in {dashboard_folder}"
 
 
 def test_remove_database(ws):
